reduce test boilerplate
