+++ virtual_labs/tests/courses/test_activate_course.py
@@ -1,110 +1,16 @@
 """Tests for the activate-course endpoint (POST /courses/{course_id}/activate)."""
 
-from typing import AsyncGenerator
 from unittest.mock import patch
 from uuid import uuid4
 
 import pytest
-import pytest_asyncio
 from httpx import AsyncClient
-from sqlalchemy import delete, select
 
-from virtual_labs.infrastructure.db.models import Course
-from virtual_labs.infrastructure.settings import settings
-from virtual_labs.shared.groups import VLAB_SERVICE_ADMIN_GROUP
-from virtual_labs.tests.utils import (
-    create_mock_lab_with_project,
-    get_headers,
-    session_context_factory,
+from virtual_labs.tests.courses.conftest import (
+    mock_admin_userinfo,
+    mock_non_admin_userinfo,
 )
-
-
-def _mock_admin_userinfo(*args, **kwargs):
-    return {"groups": [VLAB_SERVICE_ADMIN_GROUP]}
-
-
-def _mock_non_admin_userinfo(*args, **kwargs):
-    return {"groups": ["/some-other-group"]}
-
-
-async def _get_or_create_institution() -> str:
-    from virtual_labs.infrastructure.db.models import Institution
-
-    async with session_context_factory() as session:
-        result = await session.scalar(
-            select(Institution.id).where(Institution.name == "Open Brain Institute")
-        )
-        if result:
-            return str(result)
-
-        institution = Institution(
-            name="Open Brain Institute",
-            contact_email="obi-virtual-lab@openbraininstitute.org",
-        )
-        session.add(institution)
-        await session.commit()
-        await session.refresh(institution)
-        return str(institution.id)
-
-
-async def _cleanup_course(course_id: str) -> None:
-    from uuid import UUID
-
-    async with session_context_factory() as session:
-        await session.execute(delete(Course).where(Course.id == UUID(course_id)))
-        await session.commit()
-
-
-@pytest_asyncio.fixture
-async def institution_id() -> str:
-    return await _get_or_create_institution()
-
-
-@pytest_asyncio.fixture
-async def draft_course(
-    async_test_client: AsyncClient,
-    institution_id: str,
-) -> AsyncGenerator[tuple[str, str], None]:
-    """Create a course in draft status and return (course_id, vlab_id)."""
-    from uuid import UUID
-
-    from sqlalchemy import update
-
-    from virtual_labs.infrastructure.db.models import VirtualLab
-    from virtual_labs.tests.utils import cleanup_resources
-
-    lab_data, project_id = await create_mock_lab_with_project(async_test_client)
-    lab_id = lab_data["id"]
-
-    # Mark as course lab
-    async with session_context_factory() as session:
-        await session.execute(
-            update(VirtualLab)
-            .where(VirtualLab.id == UUID(lab_id))
-            .values(owner_id=settings.MULTIPLE_VLABS_ALLOWED_USER_ID)
-        )
-        await session.commit()
-
-    headers = get_headers()
-    body = {
-        "virtual_lab_id": lab_id,
-        "template_project_id": project_id,
-        "institution_id": institution_id,
-    }
-
-    with patch(
-        "virtual_labs.core.authorization.verify_service_admin.kc_auth"
-    ) as mock_kc:
-        mock_kc.userinfo.side_effect = _mock_admin_userinfo
-        response = await async_test_client.post("/courses", json=body, headers=headers)
-
-    assert response.status_code == 200
-    course_id = response.json()["data"]["id"]
-
-    yield course_id, lab_id
-
-    await _cleanup_course(course_id)
-    await cleanup_resources(async_test_client, lab_id)
+from virtual_labs.tests.utils import get_headers
 
 
 async def _set_course_dates(async_test_client: AsyncClient, course_id: str) -> None:
@@ -113,7 +19,7 @@ async def _set_course_dates(async_test_client: AsyncClient, course_id: str) -> N
     with patch(
         "virtual_labs.core.authorization.verify_service_admin.kc_auth"
     ) as mock_kc:
-        mock_kc.userinfo.side_effect = _mock_admin_userinfo
+        mock_kc.userinfo.side_effect = mock_admin_userinfo
         response = await async_test_client.patch(
             f"/courses/{course_id}",
             json={
@@ -144,7 +50,7 @@ async def test_activate_course_successfully(
     with patch(
         "virtual_labs.core.authorization.verify_service_admin.kc_auth"
     ) as mock_kc:
-        mock_kc.userinfo.side_effect = _mock_admin_userinfo
+        mock_kc.userinfo.side_effect = mock_admin_userinfo
         response = await async_test_client.post(
             f"/courses/{course_id}/activate", headers=headers
         )
@@ -172,7 +78,7 @@ async def test_activate_course_fails_when_dates_missing(
     with patch(
         "virtual_labs.core.authorization.verify_service_admin.kc_auth"
     ) as mock_kc:
-        mock_kc.userinfo.side_effect = _mock_admin_userinfo
+        mock_kc.userinfo.side_effect = mock_admin_userinfo
         response = await async_test_client.post(
             f"/courses/{course_id}/activate", headers=headers
         )
@@ -193,7 +99,7 @@ async def test_activate_course_fails_with_partial_dates(
     with patch(
         "virtual_labs.core.authorization.verify_service_admin.kc_auth"
     ) as mock_kc:
-        mock_kc.userinfo.side_effect = _mock_admin_userinfo
+        mock_kc.userinfo.side_effect = mock_admin_userinfo
         # Set only start_date
         await async_test_client.patch(
             f"/courses/{course_id}",
@@ -222,7 +128,7 @@ async def test_activate_course_fails_with_unordered_dates(
     with patch(
         "virtual_labs.core.authorization.verify_service_admin.kc_auth"
     ) as mock_kc:
-        mock_kc.userinfo.side_effect = _mock_admin_userinfo
+        mock_kc.userinfo.side_effect = mock_admin_userinfo
         # Set dates in wrong order: end_date before last_drop_date
         await async_test_client.patch(
             f"/courses/{course_id}",
@@ -254,7 +160,7 @@ async def test_activate_course_fails_when_already_active(
     with patch(
         "virtual_labs.core.authorization.verify_service_admin.kc_auth"
     ) as mock_kc:
-        mock_kc.userinfo.side_effect = _mock_admin_userinfo
+        mock_kc.userinfo.side_effect = mock_admin_userinfo
         # Activate first
         await async_test_client.post(f"/courses/{course_id}/activate", headers=headers)
         # Try to activate again
@@ -276,7 +182,7 @@ async def test_activate_course_fails_when_voided(
     with patch(
         "virtual_labs.core.authorization.verify_service_admin.kc_auth"
     ) as mock_kc:
-        mock_kc.userinfo.side_effect = _mock_admin_userinfo
+        mock_kc.userinfo.side_effect = mock_admin_userinfo
         # Void first
         await async_test_client.post(f"/courses/{course_id}/void", headers=headers)
         # Try to activate
@@ -296,7 +202,7 @@ async def test_activate_course_not_found(
     with patch(
         "virtual_labs.core.authorization.verify_service_admin.kc_auth"
     ) as mock_kc:
-        mock_kc.userinfo.side_effect = _mock_admin_userinfo
+        mock_kc.userinfo.side_effect = mock_admin_userinfo
         response = await async_test_client.post(
             f"/courses/{uuid4()}/activate", headers=headers
         )
@@ -330,7 +236,7 @@ async def test_activate_course_fails_for_non_admin(
     with patch(
         "virtual_labs.core.authorization.verify_service_admin.kc_auth"
     ) as mock_kc:
-        mock_kc.userinfo.side_effect = _mock_non_admin_userinfo
+        mock_kc.userinfo.side_effect = mock_non_admin_userinfo
         response = await async_test_client.post(
             f"/courses/{course_id}/activate", headers=headers
         )
